refactor(release): move release prompts to prompts.ts
